requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.4
h11==0.13.0
idna==3.3
install==1.3.5
lxml==4.9.1
outcome==1.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
selenium==4.4.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0
